[RFC] global/clang-format: Add .clang-format configuration for consistent formatting
This is an attempt to create and establish a common definition for
code formatting. Besides .clang-format, there don't seem to be many
good candidates for C when looking for non-proprietary open-source
solutions, so there's wasn't really much choice.

I have tried to configure everything according to the current
guidelines, to which it should come pretty close. There are a few
things for which I couldn't find any FFmpeg rules, or where it wasn't
clear and I made choices that seemed reasonable.

This is in no way intended to prescribe any formatting detail rules,
it's rather meant as an invitation to collaborate on this and ideally
end up with something "official" that allows to check and fix
formatting in CI flows or Patchwork and locally for everybody using
editors with clang-format support or using clang-format from the
command line.

Signed-off-by: softworkz <softworkz@hotmail.com>