tests/hevc: Fix concat input when running in MSYS2 shell
MSYS2 considers the colon (:) as path separator (=separating multiple
paths) and thinks it needs to convert it to a Windows
path separator (;).

Setting the MSYS2_ARG_CONV_EXCL environment variable
keeps MSYS2 from doing this replacement.

NOTE: This makes it possible to run the test from an MSYS2 shell,
but it only works when using a relative path like

--samples=../fate-suite (in configure)

or

SAMPLES=../fate-suite (in make fate)

Reviewed-by: Zhao Zhili <zhilizhao@tencent.com>
Signed-off-by: softworkz <softworkz@hotmail.com>