ffbuild/bin2c: Use zlib directly instead of gzip
Perform zlib compression directly in bin2c.c instead of
using gzip for compression, followed by bin2c to
put the file in the binary. This has the advantage of
avoiding intermediate files and also saves a few bytes
(the gzip header is longer than the zlib header).
It also avoids the hypothetical problem of the build system
supporting gzip with the zlib version used at runtime
not supporting gzip.

But the real advantages will become apparent in future commits.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>