ffbuild/bin2c: Store compressed size in-band
There are three types of users of bin2c:
a) Users that want to embed a C-string. The resources for
the fftools as well as cuda modules can be of this type
(if compression is disabled).
b) If compression is not disabled, the same users want to store
compressed data.
c) Users that just want to store arbitrary binary data.
This is currently only the videotoolbox yadif filter
for its metal shaders.

Up until now, bin2c unconditionally created a length field
for the data. Users of type a) didn't need it (and in fact actually
didn't use it); for users of type b), putting the size
of the compressed data in-band (i.e. before the actual compressed data)
is simpler. Users of type c) could currently use either an in-band
size field or an external size field.

This commit changes the cases a) and b) to no longer output
an external length field; in case b) the size is stored in-band instead
(and the data is no longer zero-terminated).
Nothing changes for c), as there is no real benefit in doing so
and I wanted to retain the ability of using external sizes
just in case.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>